Seperated common and consumer config. Using synchronous spawn to ensure task wait for the child process to be completed. Changes to generate task name to settings.
