+++ scripts/gulp-consumer-dev.js
@@ -2,12 +2,17 @@
  * Gulp task to run consumer dev server
  */
 
-const spawn = require('child_process').spawn;
+const childProcess = require('child_process').spawnSync;
 
 const runConsumerDevServer = async (done) => {
     try {
-        spawn('npm', ['run', 'dev'], { cwd: 'src/consumer/', stdio: 'inherit', shell: true })
-            .on('close', done);
+        childProcess('npm', ['run', 'dev'], { cwd: 'src/consumer/', stdio: 'inherit', shell: true }, (error) => {
+            if (error) {
+                done(error);
+                return;
+            }
+            done();
+        });
     } catch (error) {
         done(error);
     }
