+++ scripts/templates/next-sitemap.js
@@ -1,6 +1,6 @@
 module.exports = {
     siteUrl: '/*CONSUMER_URL*/',
     generateRobotsTxt: true,
-    sourceDir: './dist/consumer', //TODO: env
-    outDir: './src/consumer/public' //TODO: env
+    sourceDir: './lib', //TODO: env
+    outDir: './public' //TODO: env
 }
